remove unused code, move duplicate code together to be fixed (use interfaces)
