+++ duplicated.go
@@ -1,0 +1,42 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+)
+
+func extractProjectFromResponse(resp *http.Response, expectedCode int, listExpected bool) ([]Project, error) {
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.StatusCode != expectedCode {
+		return nil, errors.New(fmt.Sprintf("Failed with the wrong code: %v. (expected %v)\n", resp.StatusCode, expectedCode))
+	}
+
+	if listExpected {
+		return extractProjectsFromBody(body)
+	}
+
+	return extractProjectFromBody(body)
+}
+
+func extractAssetFromResponse(resp *http.Response, expectedCode int, listExpected bool) ([]Asset, error) {
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.StatusCode != expectedCode {
+		return nil, errors.New(fmt.Sprintf("Failed with the wrong code: %v. (expected %v)\n", resp.StatusCode, expectedCode))
+	}
+
+	if listExpected {
+		return extractAssetsFromBody(body)
+	}
+
+	return extractAssetFromBody(body)
+}
