+++ projects.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strconv"
@@ -126,23 +125,6 @@ func extractProjectFromBody(body []byte) ([]Project, error) {
 		return nil, err
 	}
 	return []Project{*p}, nil
-}
-
-func extractProjectFromResponse(resp *http.Response, expectedCode int, listExpected bool) ([]Project, error) {
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.StatusCode != expectedCode {
-		return nil, errors.New(fmt.Sprintf("Failed with the wrong code: %v. (expected %v)\n", resp.StatusCode, expectedCode))
-	}
-
-	if listExpected {
-		return extractProjectsFromBody(body)
-	}
-
-	return extractProjectFromBody(body)
 }
 
 func displayProjectsFromResponse(resp *http.Response, expectedCode int, listExpected bool) error {
