+++ server_interface.go
@@ -3,10 +3,7 @@ package main
 import (
 	"bytes"
 	"encoding/json"
-	"github.com/urfave/cli"
-	"io"
 	"net/http"
-	"os"
 )
 
 func Do(resource, method string, params interface{}) (*http.Response, error) {
@@ -38,22 +35,4 @@ func addAuthToHeader(hdr *http.Header, s *SlyftAuth) {
 
 func ServerURL(endpoint string) string {
 	return BackendBaseUrl + endpoint
-}
-
-// Server check functions. Perhaps not to be included in the final build.
-func PingServer(c *cli.Context) error {
-
-	url := ServerURL("/_ping")
-
-	resp, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = io.Copy(os.Stdout, resp.Body)
-	if err != nil {
-		Log.Fatal(err)
-	}
-	return nil
 }
